Move functionality of CM_Paging_List to CM_Paging_Abstract
