+++ library/CM/Paging/Abstract.php
@@ -1,6 +1,6 @@
 <?php
 
-abstract class CM_Paging_Abstract extends CM_Class_Abstract implements Iterator, Countable, CM_Cacheable {
+abstract class CM_Paging_Abstract extends CM_Class_Abstract implements Iterator, Countable, CM_Cacheable, JsonSerializable {
 
     private $_count = null;
     private $_itemsRaw = null, $_items = array(), $_itemsRawTree = null;
@@ -293,6 +293,12 @@ abstract class CM_Paging_Abstract extends CM_Class_Abstract implements Iterator,
      */
     public function setFlattenItems($state) {
         $this->_flattenItems = (bool) $state;
+    }
+
+    public function jsonSerialize() {
+        return [
+            'items' => $this->getItems()
+        ];
     }
 
     /**
