+++ library/CM/Paging/List.php
@@ -1,6 +1,6 @@
 <?php
 
-class CM_Paging_List extends CM_Paging_Abstract implements JsonSerializable {
+class CM_Paging_List extends CM_Paging_Abstract {
 
     /**
      * @param CM_PagingSource_Array|array $source
@@ -14,11 +14,5 @@ class CM_Paging_List extends CM_Paging_Abstract implements JsonSerializable {
             throw new CM_Exception_Invalid('CM_Paging_List should be instantiated with either an array or CM_PagingSource_Array instance.');
         }
         parent::__construct($source);
-    }
-
-    public function jsonSerialize() {
-        return [
-            'list' => $this->getItems()
-        ];
     }
 }
